requirements.txt
# script.py — real-time Speechmatics + .env
speechmatics-rt==1.0.0
python-dotenv>=1.0,<2
# явная зависимость: speechmatics-rt импортирует typing_extensions.Self
typing-extensions>=4.12.0